+++ Config/routes.php
@@ -56,6 +56,7 @@ Router::connect('/post/*',			array('controller' => 'posts', 'action' => 'view',	
 //Profiles controller
 Router::connect('/password/request',		array('controller' => 'profiles', 'action' => 'request_new_password',	'plugin' => 'me_cms'));
 Router::connect('/password/:id/:token',	array('controller' => 'profiles', 'action' => 'reset_password',			'plugin' => 'me_cms'), array('pass' => array('id', 'token')));
+Router::connect('/signup',				array('controller' => 'profiles', 'action' => 'signup',					'plugin' => 'me_cms'));
 
 //System controller
 Router::connect('/offline', array('controller' => 'systems', 'action' => 'offline', 'plugin' => 'me_cms'));